Separate into own files (#1)
* Separate into own files

* Add badge
