+++ lib/EvtListener/EvtListener.cpp
@@ -1,0 +1,26 @@
+#include "EvtListener.h"
+
+void EvtListener::setupListener()
+{
+}
+
+bool EvtListener::isEventTriggered()
+{
+  return enabled;
+}
+
+bool EvtListener::performTriggerAction(EvtContext *ctx)
+{
+  return (*triggerAction)(this, ctx);
+}
+
+void EvtListener::disable()
+{
+  this->enabled = false;
+}
+  
+void EvtListener::enable()
+{
+  this->enabled = true;
+  this->setupListener();
+}
