+++ platformio.ini
@@ -1,0 +1,27 @@
+; PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; https://docs.platformio.org/page/projectconf.html
+
+[platformio]
+default_envs = uno
+
+[env:uno]
+platform = atmelavr
+board = uno
+framework = arduino
+monitor_filters = time, default
+monitor_speed = 9600
+
+[env:native]
+platform = native
+
+[env:debug]
+platform = native
+build_type = debug
+debug_test = test_display
